fix(API): don't allow running pipelines to be deleted
It can lead to interesting race conditions.

+++ orca-web/src/main/groovy/com/netflix/spinnaker/orca/controllers/TaskController.groovy
@@ -16,15 +16,15 @@
 
 package com.netflix.spinnaker.orca.controllers
 
-import com.netflix.spinnaker.orca.model.OrchestrationViewModel
-
 import java.time.Clock
 import com.netflix.spinnaker.orca.ExecutionStatus
 import com.netflix.spinnaker.orca.front50.Front50Service
+import com.netflix.spinnaker.orca.model.OrchestrationViewModel
 import com.netflix.spinnaker.orca.pipeline.ExecutionRunner
 import com.netflix.spinnaker.orca.pipeline.PipelineStartTracker
 import com.netflix.spinnaker.orca.pipeline.StageDefinitionBuilder
 import com.netflix.spinnaker.orca.pipeline.model.Execution
+import com.netflix.spinnaker.orca.pipeline.model.Execution.ExecutionType
 import com.netflix.spinnaker.orca.pipeline.model.Stage
 import com.netflix.spinnaker.orca.pipeline.persistence.ExecutionRepository
 import com.netflix.spinnaker.orca.pipeline.util.ContextParameterProcessor
@@ -180,7 +180,14 @@ class TaskController {
   @PreAuthorize("hasPermission(this.getPipeline(#id)?.application, 'APPLICATION', 'WRITE')")
   @RequestMapping(value = "/pipelines/{id}", method = RequestMethod.DELETE)
   void deletePipeline(@PathVariable String id) {
-    executionRepository.delete(PIPELINE, id)
+    executionRepository.retrieve(PIPELINE, id).with {
+      if (it.status.complete) {
+        executionRepository.delete(PIPELINE, id)
+      } else {
+        log.warn("Not deleting $PIPELINE $id as it is $it.status")
+        throw new CannotDeleteRunningExecution(PIPELINE, id)
+      }
+    }
   }
 
   @PreAuthorize("hasPermission(this.getPipeline(#id)?.application, 'APPLICATION', 'WRITE')")
@@ -379,4 +386,11 @@ class TaskController {
   @InheritConstructors
   @ResponseStatus(HttpStatus.NOT_IMPLEMENTED)
   private static class FeatureNotEnabledException extends RuntimeException {}
+
+  @ResponseStatus(HttpStatus.CONFLICT)
+  private static class CannotDeleteRunningExecution extends RuntimeException {
+    CannotDeleteRunningExecution(ExecutionType type, String id) {
+      super("Cannot delete a running $type, please cancel it first.")
+    }
+  }
 }
